[API] Check not only _fail but also _criticalFail when determining PASS
In some cases, the report function is used to 'literally' report errors
that occur in test configuration procedure before terminating the test.
For the convenience, this patch makes _criticalFail be also checked when
determining PASS/FAIL of the tests. Therefore, to use for reporting
critical errors to users or systems, only one variable, _criticalFail,
is enough instead of setting both of them.

Signed-off-by: Wook Song <wook16.song@samsung.com>